Change Panel content to be a str.
This was necessary b/c otherwise, passing in e.g. ds.P meant that the para inside the panel had "govuk-body" class. This made the text black whereas the text inside a panel should be white.

This is not a perfect solution b/c sometimes, content would need some tags. I have included an example on how to do that with ds.Safe.